ModelAPI/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
requests==2.32.3
Pillow==10.4.0
numpy==1.26.4
tensorflow==2.16.1
transformers==4.44.2
accelerate==0.33.0